added dedicated enpoint path for alerts config
