+++ backend/endpoints/api/v2/__init__.py
@@ -17,3 +17,4 @@ from . import internaluser
 from . import role
 from . import git
 from . import notifications
+from . import alerts
