+++ backend/endpoints/api/v2/db.py
@@ -17,14 +17,11 @@ import urllib.parse
 from datetime import datetime
 from typing import Any, List, Tuple, Dict
 from aiohttp import web, web_request
-
 from dal.backup import BackupManager, RestoreManager
 from dal.data import WorkspaceManager
 from dal.models.scopestree import scopes
 from dal.models.model import Model
 from dal.models.user import User
-from dal.models.var import Var
-
 from backend.http import WebAppManager
 from backend.endpoints.api.v2.base import BaseWebApp
 
@@ -772,63 +769,6 @@ async def rebuild_indexes(request: web.Request):
     return {"status": "workspace indexes rebuild started"}
 
 
-async def _set_alerts_config(data: dict):
-    """
-    Set the alerts config
-    """
-    var_global = Var("global")
-    setattr(var_global, "alertsConfig", data)
-
-
-async def set_alerts_config(request: web.Request):
-    data = await request.json()
-    if "alerts" not in data:
-        raise web.HTTPBadRequest(reason="alerts not in data")
-
-    alerts = data["alerts"]
-    _check_user_permission(request, "EmailsAlertsConfig", "update")
-    alertsConfig = _get_alerts_config()
-    alertsConfig["alerts"] = alerts
-    _set_alerts_config(alertsConfig)
-    return web.json_response(
-        alertsConfig,
-        headers={"Server": "Movai-server"},
-    )
-
-
-async def set_alerts_emails(request: web.Request):
-    data = await request.json()
-    if "emails" not in data:
-        raise web.HTTPBadRequest(reason="\"emails\" not in data")
-
-    recipients = data["emails"]
-    _check_user_permission(request, "EmailsAlertsRecipients", "update")
-
-    alertsConfig = _get_alerts_config()
-    alertsConfig["emails"] = recipients
-    _set_alerts_config(alertsConfig)
-    return web.json_response(
-        alertsConfig,
-        headers={"Server": "Movai-server"},
-    )
-
-
-def _get_alerts_config() -> Dict[str, Any]:
-    return Var("global").get("alertsConfig")
-
-
-async def get_alerts_emails(request: web.Request) -> web.json_response:
-    _check_user_permission(request, "AlertsRecipients", "read")
-    alertsConfig = _get_alerts_config()
-    return web.json_response(alertsConfig["emails"], headers={"Server": "Movai-server"})
-
-
-async def get_alerts_config(request: web.Request):
-    _check_user_permission(request, "AlertsConfig", "read")
-    alertsConfig = _get_alerts_config()
-    return web.json_response(alertsConfig["alerts"], headers={"Server": "Movai-server"})
-
-
 class DatabaseAPI(BaseWebApp):
     """Web application for serving as the database api."""
 
@@ -851,10 +791,6 @@ class DatabaseAPI(BaseWebApp):
             web.get(r"/restore", get_restore_jobs_list),
             web.get(r"/restore/{job_id}", get_restore_state),
             web.get(r"/restore/{job_id}/log", get_restore_log),
-            web.post(r"/alerts/emails", set_alerts_emails),
-            web.post(r"/alerts/config", set_alerts_config),
-            web.get(r"/alerts/emails", get_alerts_emails),
-            web.get(r"/alerts/config", get_alerts_config),
             web.post(r"/restore/clean", start_restore_clean),
             web.post(r"/{workspace}/rebuild-indexes", rebuild_indexes),
             web.post(r"/{workspace}", create_workspace),
